// src/store/salesTax/salesTax.enum.ts
enum SalesTaxActionTypeEnum {
	GET_STATES_SUCCESS = "GET_STATES_SUCCESS",
	GET_STATES_PENDING = "GET_STATES_PENDING",
	GET_STATES_FAILED = "GET_STATES_FAILED",

	ADD_STATE_SUCCESS = "ADD_STATE_SUCCESS",
	ADD_STATE_FAILED = "ADD_STATE_FAILED",
	ADD_STATE_PENDING = "ADD_STATE_PENDING",

	UPDATE_STATE_SUCCESS = "UPDATE_STATE_SUCCESS",
	UPDATE_STATE_FAILED = "UPDATE_STATE_FAILED",
	UPDATE_STATE_PENDING = "UPDATE_STATE_PENDING",

	DELETE_STATE_SUCCESS = "DELETE_STATE_SUCCESS",
	DELETE_STATE_FAILED = "DELETE_STATE_FAILED",
	DELETE_STATE_PENDING = "DELETE_STATE_PENDING",

	GET_COUNTIES_SUCCESS = "GET_COUNTIES_SUCCESS",
	GET_COUNTIES_PENDING = "GET_COUNTIES_PENDING",
	GET_COUNTIES_FAILED = "GET_COUNTIES_FAILED",

	ADD_COUNTIES_SUCCESS = "ADD_COUNTIES_SUCCESS",
	ADD_COUNTIES_PENDING = "ADD_COUNTIES_PENDING",
	ADD_COUNTIES_FAILED = "ADD_COUNTIES_FAILED",

	UPDATE_COUNTIES_SUCCESS = "UPDATE_COUNTIES_SUCCESS",
	UPDATE_COUNTIES_PENDING = "UPDATE_COUNTIES_PENDING",
	UPDATE_COUNTIES_FAILED = "UPDATE_COUNTIES_FAILED",

	DELETE_COUNTIES_SUCCESS = "DELETE_COUNTIES_SUCCESS",
	DELETE_COUNTIES_FAILED = "DELETE_COUNTIES_FAILED",
	DELETE_COUNTIES_PENDING = "DELETE_COUNTIES_PENDING",

	GET_ZIPCODES_SUCCESS = "GET_ZIPCODES_SUCCESS",
	GET_ZIPCODES_PENDING = "GET_ZIPCODES_PENDING",
	GET_ZIPCODES_FAILED = "GET_ZIPCODES_FAILED",

	ADD_ZIPCODES_SUCCESS = "ADD_ZIPCODES_SUCCESS",
	ADD_ZIPCODES_PENDING = "ADD_ZIPCODES_PENDING",
	ADD_ZIPCODES_FAILED = "ADD_ZIPCODES_FAILED",

	UPDATE_ZIPCODE_SUCCESS = "UPDATE_ZIPCODE_SUCCESS",
	UPDATE_ZIPCODE_PENDING = "UPDATE_ZIPCODE_PENDING",
	UPDATE_ZIPCODE_FAILED = "UPDATE_ZIPCODE_FAILED",

	DELETE_ZIPCODE_SUCCESS = "DELETE_ZIPCODE_SUCCESS",
	DELETE_ZIPCODE_PENDING = "DELETE_ZIPCODE_PENDING",
	DELETE_ZIPCODE_FAILED = "DELETE_ZIPCODE_FAILED",
}

export default SalesTaxActionTypeEnum;